[fix] int 범위를 넘어가는 이익의 수익률 계산
int범위를 넘어가는 이익이 들어올 때 수익률 계산 에러 해결
double형으로 교체하였다.

+++ src/main/java/lotto/domain/number/PurchaseNumber.java
@@ -35,7 +35,7 @@ public class PurchaseNumber {
         return purchaseNumber;
     }
 
-    public int getPurchaseMoney() {
+    public double getPurchaseMoney() {
         return purchaseNumber * 1000;
     }
 }
